backend/requirements.txt
Flask>=2.0
google-generativeai>=0.5.0
Flask-SQLAlchemy>=2.5
Flask-CORS>=3.0.10 # Added Flask-CORS
requests>=2.25.0
opencv-python>=4.5.0
google-cloud-storage>=1.31.0
google-cloud-aiplatform>=1.38.0 # For Vertex AI integration
# mediapy
python-dotenv>=0.15.0 # For loading .env files
# Add other specific versions if known or required
